fix: ci intValue bug

+++ xcloud-iam-core/src/main/java/com/wl4g/iam/service/impl/MenuServiceImpl.java
@@ -66,7 +66,7 @@ public class MenuServiceImpl implements MenuService {
 
 	private Menu getParent(List<Menu> menus, Long parentId) {
 		for (Menu menu : menus) {
-			if (parentId != null && menu.getId() != null && menu.getId().intValue() == parentId.intValue()) {
+			if (parentId != null && menu.getId() != null && menu.getId().longValue() == parentId.longValue()) {
 				return menu;
 			}
 		}
